feat: 프로모션 로직 구현

+++ src/main/kotlin/store/model/Product.kt
@@ -9,7 +9,7 @@ class Product(val productInfo: ProductInfo, val quantity: Quantity, val promotio
             val promotionList = PromotionList().getPromotionList()
 
             val productInfo = ProductInfo(productList.first().toString(), productList[1].toInt())
-            val quantity = Quantity(productList[2])
+            val quantity = Quantity(productList[2].toInt())
             val promotion = promotionList.find { it.promotion == (productList[3]) }!!
             return Product(productInfo, quantity, promotion)
         }
